requirements.txt
streamlit>=1.35.0
plotly>=5.17.0
pandas>=2.0.3
numpy>=1.24.3
requests>=2.31.0
aiohttp>=3.9.0
wordcloud>=1.9.2
matplotlib>=3.7.1
seaborn>=0.12.2
python-dotenv>=1.0.0
certifi~=2025.6.15